backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
requests
python-dotenv
# Optional / advanced
PySpice
# google-genai (or google-genai equivalent) - install when ready
# google-genai
